Worked On SignIn/SignOut Button
Attempting to implement a navigation menu button that alternates between Sign In and Sign Out depending on the users session state. Currently the client side Sign Out button doesn't work and when the user signs in the button doesn't update to Sign Out unless if the user is redirected to the home page.